Greatly simplify lime.app.Event
Instead of generating a full class per each listener argument combination we now just redirect to one of 4 possible Event subclasses for different number of arguments, specifying argument types as type parameters.

This saves quite some generated code, which is of course good. In general I think we should get rid of Event in most of the places and use direct function calls, because we don't need this kind of modularity here.

Oh, also removed the "listen once" functionality, because that's just the abuse of the observer pattern idea. Use a callback if you need to listen once.