remove commented out (and most probably defunct) filter shader stuff
we *do want* shader-based filters in future so we might want to look through the git history for shader sources, but we surely won't use this as is, because this was never integrated in the original openfl rendering pipeline, and now when we changed everything with the batcher and stuff it's even more incompatible.

also removing this gives some space for Shader-related code refactoring \o/

also remove ShaderFilter. we should also remove the whole Shader/ShaderData/etc. public API and move it in the `_internal`, because we don't even support shader-from-bytecode loading. and while it's technically possible on a lot of targets (load bytecode -> translate to glsl -> compile), i'm 100% sure nobody will ever need this, so the crazy high cost of trying to integrate this with everything else is just not worth it.