Rework Bitmap+BitmapData dirty checking and remove __enterFrame
Now instead of traversing the whole tree to check and mark Bitmaps with modified BitmapDatas, we explicitly check a subtree we're trying to render. We also only do this for cacheAsBitmap'd object, because for the GL rendering we don't care about the __renderDirty flags at all (tho we do the modification check in Bitmap.__getBatchQuad, although I'm not sure if this is any better than just always calling getTexture, because it will check the version again).