Port a quick composed character input support from upstream OpenFL
Here we simply ignore input while composition is active and process the input after composition ends. It's not ideal because TextField won't show the current composition, but at least one will be able to enter composed characters.

Later we can work on communicating composition updates to TextField so it can display it nicely.